Rename Generate to GenerateFromPlan
